requirements.txt
streamlit
pandas
pytesseract
Pillow
langchain
langchain-community
faiss-cpu
ollama
docx2txt
pypdf
unstructured
openpyxl